# Conflicts:
#	src/main/java/com/ohgiraffers/ForepeApplication.java
#	src/main/java/com/ohgiraffers/forepeproject/login/command/application/controller/LoginController.java
#	src/main/java/com/ohgiraffers/forepeproject/login/command/application/service/LoginService.java
#	src/main/java/com/ohgiraffers/forepeproject/login/command/domain/repository/LoginRepository.java
#	src/main/java/com/ohgiraffers/forepeproject/member/command/application/controller/MemberController.java
#	src/main/java/com/ohgiraffers/forepeproject/member/command/application/service/MemberService.java
#	src/main/java/com/ohgiraffers/forepeproject/member/query/domain/repository/MemberRepository.java
#	src/main/resources/config/application.yml
#	src/main/resources/static/html/deleteaccount.html
#	src/main/resources/static/html/editnickname.html
#	src/main/resources/static/html/feedback.html
#	src/main/resources/static/html/index.html
#	src/main/resources/static/html/lactoovo.html
#	src/main/resources/static/html/login.html
#	src/main/resources/static/html/mytree.html
#	src/main/resources/static/html/pescopolo.html
#	src/main/resources/static/html/recipeReview.html
#	src/main/resources/static/html/vegan.html
#	src/main/resources/static/html/view.html
#	src/main/resources/static/html/write.html

+++ src/main/java/com/ohgiraffers/forepeproject/member/command/application/dto/MemberDTO.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @ToString
 public class MemberDTO {
-    private int MemberNum;
+    private long MemberNum;
     private String MemberName;
     private String socialLogin;
     private long socialId;
